backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
requests==2.32.3
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
pinecone>=4.0.0
python-dotenv==1.0.1
numpy==1.26.4